+++ platform/RELEASES.md
@@ -59,3 +59,13 @@ Based on:
 - [typescript v1.3.0] platform
 ### Releases
 - [NPM v1.3.0] https://www.npmjs.com/package/@codat/platform/v/1.3.0 - platform
+
+## 2023-10-30 01:28:16
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Platform.yaml
+- Speakeasy CLI 1.109.0 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.4.0] platform
+### Releases
+- [NPM v1.4.0] https://www.npmjs.com/package/@codat/platform/v/1.4.0 - platform